+++ Tests/Fixtures/ActionArgumentsBundle/Controller/NullableArgumentsController.php
@@ -5,8 +5,8 @@ namespace Tests\Fixtures\ActionArgumentsBundle\Controller;
 use Psr\Http\Message\MessageInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
 
 /**
  * @Route("/nullable-arguments")