+++ Tests/Fixtures/FooBundle/Controller/InvokableClassLevelController.php
@@ -11,11 +11,10 @@
 
 namespace Tests\Fixtures\FooBundle\Controller;
 
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
+use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route(service="test.invokable_class_level.predefined")
  * @Template("@Foo/invokable/predefined.html.twig")
  */
 class InvokableClassLevelController