+++ Tests/Fixtures/FooBundle/Controller/InvokableController.php
@@ -11,12 +11,9 @@
 
 namespace Tests\Fixtures\FooBundle\Controller;
 
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
+use Symfony\Component\Routing\Annotation\Route;
 
-/**
- * @Route(service="test.invokable.predefined")
- */
 class InvokableController
 {
     /**