+++ Tests/Fixtures/FooBundle/Controller/MultipleActionsClassLevelTemplateController.php
@@ -11,9 +11,9 @@
 
 namespace Tests\Fixtures\FooBundle\Controller;
 
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\Routing\Annotation\Route;
 
 /**
  * @Template("@Foo/invokable/predefined.html.twig")